Address editor rendering review feedback
