+++ src/components/video-editor/canvas/composables/tests/use-camera-zoom-test-types.ts
@@ -1,0 +1,3 @@
+import type { compositeIsolatedMotionBlurSample } from '../../../zoom/zoom-motion-blur-compositor';
+
+export type CompositeMotionBlurOptions = Parameters<typeof compositeIsolatedMotionBlurSample>[0];
